backend/requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-jose
argon2-cffi